chore: refactor frontend for projects (project table and modals)
