+++ FrontEndApp/v1-react/src/mocks/resolvers/createProjectsResolver.js
@@ -15,8 +15,12 @@ const mockData = {
 };
 
 export const createMockProject = (req, res, ctx) => {
+    let status = 200;
+    let success = True;
+    let message = "Project Created Successfully";
+
     return res(
-        ctx.status(200),
-        ctx.json({ projects: mockResponse }),
-    )
+        ctx.status(status),
+        ctx.json({ "success": success, "message": message })
+    );
 }
